Extract a new dependency module, `@bayou/deps-util`.
