+++ local-modules/@bayou/deps-util/README.md
@@ -1,0 +1,5 @@
+@bayou/deps-util
+================
+
+This module just serves as a single location to hold external dependencies
+for general utility functionality.
