+++ local-modules/@bayou/deps-util/index.js
@@ -1,0 +1,7 @@
+// Copyright 2016-2018 the Bayou Authors (Dan Bornstein et alia).
+// Licensed AS IS and WITHOUT WARRANTY under the Apache License,
+// Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
+
+// This module serves only to hold common external dependencies. There is
+// nothing to export.
+export {};
