+++ local-modules/@bayou/deps-util/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@bayou/deps-util",
+  "version": "1.0.0",
+
+  "dependencies": {
+    "lodash": "^4.17.4"
+  }
+}
