+++ local-modules/@bayou/state-machine/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "state-machine",
+  "name": "@bayou/state-machine",
   "version": "1.0.0",
 
   "dependencies": {
